third switch case must be constant value

+++ src/hash.c
@@ -1,14 +1,14 @@
 #include "../head/SERVER.h"
 unsigned int DJBHash(char* str, unsigned int len)
 {
-   unsigned int hash = 5381;
-   unsigned int i    = 0;
+    unsigned int hash = 5381;
+    unsigned int i    = 0;
 
-   for(i = 0; i < len; str++, i++)
-   {
-      hash = ((hash << 5) + hash) + (*str);
-   }
+    for(i = 0; i < len; str++, i++)
+    {
+        hash = ((hash << 5) + hash) + (*str);
+    }
 
-   return hash%193400000;
+    return (int)(hash%193400000);
 }
 /* End Of DJB Hash Function */
